etc/bashrc: Do not add our bin to path if already present
By the name, this file is clearly intended to be invoked from a user's
~/.bashrc. However, bashrc is re-invoked for subshells that inherit
a parent shell's environment and thus must be idempotent about
changing environment variables such as $PATH or directories may be
added to $PATH multiple times.

This adds $PATH_ROOT/bin to $PATH only if it's not already in $PATH to
avoid adding multiple copies.

Also, we no longer export $PATH because it's not necessary; changing
already-exported variables changes the environment. (And if it's not
exported to the environment for some bizzarre reason, we should
probably not override that decision.)